+++ R/cv_vim.R
@@ -7,15 +7,15 @@
 #'
 #' @inheritParams vim
 #' @param cross_fitted_f1 the predicted values on validation data from a
-#'   flexible estimation technique regressing Y on X in the training data;
-#'   a list of length V, where each object is a set of predictions on the
-#'   validation data. If sample-splitting is requested, then these must
+#'   flexible estimation technique regressing Y on X in the training data provided as
+#'   a list of length V, where each element in the list is a set of predictions on the
+#'   corresponding validation data fold. If sample-splitting is requested, then these must
 #'   be estimated specially; see Details.
 #' @param cross_fitted_f2 the predicted values on validation data from a
 #'   flexible estimation technique regressing either (a) the fitted values in
-#'   \code{cross_fitted_f1}, or (b) Y, on X withholding the columns in \code{indx};
-#'   a list of length V, where each object is a set of predictions on the
-#'   validation data. If sample-splitting is requested, then these must
+#'   \code{cross_fitted_f1}, or (b) Y, on X withholding the columns in \code{indx}
+#'   provided as a list of length V, where each object is a set of predictions on the
+#'   corresponding validation data fold. If sample-splitting is requested, then these must
 #'   be estimated specially; see Details.
 #' @param f1 the fitted values from a flexible estimation technique
 #'   regressing Y on X. If sample-splitting is requested, then these must be