+++ NEWS.md
@@ -9,6 +9,8 @@
 * Allow different types of bootstrap interval (e.g., percentile) to be computed
 * More precise documentation for `Z` in coarsened-data settings; allow case-insensitive specification of covariate names/positions when creating `Z`
 * `V` defaults to 5 if no cross-fitting folds are specified externally
+* More precise documentation for `cross_fitted_f1` and `cross_fitted_f2` in `cv_vim`
+* Added an input check for `cross_fitted_f1` and `cross_fitted_f2`; these must be lists
 
 # vimp 2.2.5